Separate some classes into their own files.
