+++ Lib/Objects/FileSystemMonitorResult.cs
@@ -1,0 +1,14 @@
+﻿using System;
+namespace AttackSurfaceAnalyzer.Objects
+{
+    public class FileSystemMonitorResult
+    {
+        public FileSystemEventArgs evt { get; set; }
+        public NotifyFilters filter { get; set; }
+
+        public FileSystemMonitorResult(FileSystemEventArgs evtIn)
+        {
+            evt = evtIn;
+        }
+    }
+}
