+++ Lib/Objects/Rule.cs
@@ -22,18 +22,4 @@ namespace AttackSurfaceAnalyzer.Objects
             this.Name = Name;
         }
     }
-
-    public class Clause
-    {
-        public string Field { get; set; }
-        public OPERATION Operation { get; set; }
-        public string? Label { get; set; }
-        public List<string>? Data { get; set; }
-        public List<KeyValuePair<string, string>>? DictData { get; set; }
-        public Clause(string Field, OPERATION Operation)
-        {
-            this.Field = Field;
-            this.Operation = Operation;
-        }
-    }
 }
