+++ Lib/Objects/Settings.cs
@@ -8,6 +8,5 @@ namespace AttackSurfaceAnalyzer.Objects
         public int ShardingFactor { get; set; }
         public bool TelemetryEnabled { get; set; }
         public int SchemaVersion { get; set; }
-
     }
 }
